tests folder + renaming tests
